Add producer keychain JWK key model
