+++ packages/models/src/authorization/producerKeychain.ts
@@ -5,7 +5,7 @@ import {
   TenantId,
   UserId,
 } from "../brandedIds.js";
-import { Key } from "./key.js";
+import { JWKKey, Key } from "./key.js";
 
 export const ProducerKeychainKey = Key.extend({
   producerKeychainId: ProducerKeychainId,
@@ -23,3 +23,9 @@ export const ProducerKeychain = z.object({
   keys: z.array(ProducerKeychainKey),
 });
 export type ProducerKeychain = z.infer<typeof ProducerKeychain>;
+
+export const ProducerKeychainJWKKey = JWKKey.extend({
+  producerKeychainId: ProducerKeychainId,
+});
+
+export type ProducerKeychainJWKKey = z.infer<typeof ProducerKeychainJWKKey>;
